+++ ls/server/src/main/kotlin/org/opencds/cqf/cql/ls/server/command/CliCommand.kt
@@ -1,7 +1,0 @@
-package org.opencds.cqf.cql.ls.server.command
-
-import org.opencds.cqf.cql.ls.server.manager.IgContextManager
-import picocli.CommandLine.Command
-
-@Command(subcommands = [CqlCommand::class])
-class CliCommand(val igContextManager: IgContextManager)